Add repo + endpoint

+++ CasbinMinimalApi/Application/Repositories/IScissorsGenericRepository.cs
@@ -5,7 +5,7 @@ namespace CasbinMinimalApi.Application.Repositories;
 public interface IScissorsGenericRepository<T> where T : Entity
 {
   Task<IEnumerable<T>> GetAllAsync();
-  Task<T?> GetByIdAsync(int id);
+  Task<T?> GetByIdAsync(long id);
   Task AddAsync(T entity);
   void Update(T entity);
   void Delete(T entity);
